Move mergeEventHandlers to object-to-react
