+++ packages/transform-vdom/src/index.tsx
@@ -1,8 +1,7 @@
 import * as React from "react";
 import { cloneDeep } from "lodash";
 
-import { objectToReactElement, VDOMEl, Attributes } from "./object-to-react";
-import { serializeEvent, SerializedEvent } from "./event-to-object";
+import { objectToReactElement, VDOMEl, Attributes, SerializedEvent } from "./object-to-react";
 
 interface Props {
   mediaType: "application/vdom.v1+json";
@@ -11,7 +10,7 @@ interface Props {
 }
 
 // Provide object-to-react as an available helper on the library
-export { objectToReactElement, serializeEvent, VDOMEl, Attributes, SerializedEvent };
+export { objectToReactElement, VDOMEl, Attributes, SerializedEvent };
 
 const mediaType = "application/vdom.v1+json";
 
@@ -24,8 +23,9 @@ export default class VDOM extends React.PureComponent<Props> {
 
   render(): React.ReactElement<any> {
     try {
-      const obj = this.mergeEventHandlers(cloneDeep(this.props.data));
-      return objectToReactElement(obj);
+      // objectToReactElement is mutatitve so we'll clone our object
+      var obj = cloneDeep(this.props.data);
+      return objectToReactElement(obj, this.props.onVDOMEvent);
     } catch (err) {
       return (
         <React.Fragment>
@@ -46,21 +46,4 @@ export default class VDOM extends React.PureComponent<Props> {
       );
     }
   }
-
-  // Merge event handlers (if any) into attributes as callback functions that
-  // serialize the event object and call `this.props.onVDOMEvent` with the
-  // comm target name and serialized event so that the Jupyter client can
-  // send a comm message for the kernel to handle the event.
-  mergeEventHandlers = (obj: VDOMEl): VDOMEl => {
-    if (obj.eventHandlers) {
-      for (let eventType in obj.eventHandlers) {
-        const targetName = obj.eventHandlers[eventType];
-        obj.attributes[eventType] = (event: React.SyntheticEvent<any>) => {
-          const serializedEvent = serializeEvent(event);
-          this.props.onVDOMEvent(targetName, serializedEvent);
-        };
-      }
-    }
-    return obj;
-  };
 }
