+++ packages/transform-vdom/src/object-to-react.ts
@@ -32,6 +32,9 @@
  */
 
 import * as React from "react";
+import { serializeEvent, SerializedEvent } from "./event-to-object";
+
+export { SerializedEvent } from "./event-to-object";
 
 export interface Attributes {
   [key: string]: any;
@@ -58,13 +61,59 @@ export interface VDOMEl {
  * @param  {Object}       obj - The element object.
  * @return {ReactElement}
  */
-export function objectToReactElement(obj: VDOMEl): React.ReactElement<any> {
+export function objectToReactElement(
+  obj: VDOMEl,
+  onVDOMEvent: (targetName: string, event: SerializedEvent<any>) => void
+): React.ReactElement<any> {
+  /**
+   * Convert an array of items to React children.
+   *
+   * @param  {Array} arr - The array.
+   * @return {Array}     - The array of mixed values.
+   */
+  function arrayToReactChildren(
+    arr: Array<VDOMEl>
+  ): React.ReactNodeArray {
+    let result = [];
+
+    // iterate through the `children`
+    for (let i = 0, len = arr.length; i < len; i++) {
+      // child can have mixed values: text, React element, or array
+      const item = arr[i];
+      if (item === null) {
+        continue;
+      } else if (Array.isArray(item)) {
+        result.push(arrayToReactChildren(item));
+      } else if (typeof item === "string") {
+        result.push(item);
+      } else if (typeof item === "object") {
+        // Create a new object so that if we have to set the key, we are not
+        // mutating the original object
+        const keyedItem = {
+          tagName: item.tagName,
+          attributes: item.attributes,
+          children: item.children,
+          key: i
+        };
+        if (item.attributes && item.attributes.key) {
+          keyedItem.key = item.attributes.key;
+        }
+        result.push(objectToReactElement(keyedItem, onVDOMEvent));
+      } else {
+        throw new Error(`invalid vdom child: "${item}"`);
+      }
+    }
+
+    return result;
+  }
+
   // Pack args for React.createElement
   let args = [];
 
   if (!obj.tagName || typeof obj.tagName !== "string") {
     throw new Error(`Invalid tagName on ${JSON.stringify(obj, null, 2)}`);
   }
+
   if (
     !obj.attributes ||
     Array.isArray(obj.attributes) ||
@@ -86,6 +135,22 @@ export function objectToReactElement(obj: VDOMEl): React.ReactElement<any> {
     );
   }
 
+  // Add event handlers to attributes.
+  // Replace event handlers values (target name) with callback functions that
+  // serialize the event object and call `onVDOMEvent` with the comm target name
+  // and serialized event.
+  // `onVDOMEvent` will send a comm message to a comm channel of target name
+  // with a body of serialized event and vdom on kernel will handle the event.
+  if (obj.eventHandlers) {
+    for (let eventType in obj.eventHandlers) {
+      const targetName = obj.eventHandlers[eventType];
+      obj.attributes[eventType] = (event: React.SyntheticEvent<any>) => {
+        const serializedEvent = serializeEvent(event);
+        onVDOMEvent(targetName, serializedEvent);
+      };
+    }
+  }
+
   // `React.createElement` 1st argument: type
   args[0] = obj.tagName;
   args[1] = obj.attributes;
@@ -102,7 +167,7 @@ export function objectToReactElement(obj: VDOMEl): React.ReactElement<any> {
     } else if (typeof children === "string") {
       args[2] = children;
     } else if (typeof children === "object") {
-      args[2] = objectToReactElement(children as VDOMEl);
+      args[2] = objectToReactElement(children as VDOMEl, onVDOMEvent);
     } else {
       throw new Error(
         "children of a vdom element must be a string, object, null, or array of vdom nodes"
@@ -112,44 +177,4 @@ export function objectToReactElement(obj: VDOMEl): React.ReactElement<any> {
 
   // $FlowFixMe: React
   return React.createElement.apply({}, args);
-}
-
-/**
- * Convert an array of items to React children.
- *
- * @param  {Array} arr - The array.
- * @return {Array}     - The array of mixed values.
- */
-export function arrayToReactChildren(arr: Array<VDOMEl>): React.ReactNodeArray {
-  let result = [];
-
-  // iterate through the `children`
-  for (let i = 0, len = arr.length; i < len; i++) {
-    // child can have mixed values: text, React element, or array
-    const item = arr[i];
-    if (item === null) {
-      continue;
-    } else if (Array.isArray(item)) {
-      result.push(arrayToReactChildren(item));
-    } else if (typeof item === "string") {
-      result.push(item);
-    } else if (typeof item === "object") {
-      // Create a new object so that if we have to set the key, we are not
-      // mutating the original object
-      const keyedItem = {
-        tagName: item.tagName,
-        attributes: item.attributes,
-        children: item.children,
-        key: i
-      };
-      if (item.attributes && item.attributes.key) {
-        keyedItem.key = item.attributes.key;
-      }
-      result.push(objectToReactElement(keyedItem));
-    } else {
-      throw new Error(`invalid vdom child: "${item}"`);
-    }
-  }
-
-  return result;
 }
